First stab at fixing Issue 223.
We keep around mtimes from one iteration of DependencyWatcher to another, so
that if something is (eg) modified during a slow NPM update during the bundler,
another restart will happen.

It still doesn't quite work. Two problems:

(a) It still doesn't deal with the fact that on OSX HFS+, mtimes have a 1-second
    resolution. So if a file is written twice within a second, we might read the first
    version and never notice the second version. (This can be slightly improved by
    also reading the size, and fully fixed by storing hashes instead of mtimes.)

(b) Now that the initial scan can cause a restart, we can run into an issue
    where restart_server() is called (from on_change) before we even get to
    start_server!